refactor: split schedule-updated handler into tasks/ convention
